Issue #3333: Implement BIOVIA 2009 valence mode across loaders and models for consistent valence calculations
