+++ core/indigo-core/molecule/src/cml_loader.cpp
@@ -46,7 +46,7 @@ CmlLoader::CmlLoader(Scanner& scanner)
     _scanner = &scanner;
     _handle = 0;
     ignore_bad_valence = false;
-    valence_mode = ValenceMode::DEFAULT;
+    valence_mode = ValenceMode::BIOVIA_2009;
 }
 
 CmlLoader::CmlLoader(XMLHandle& handle)
@@ -54,7 +54,7 @@ CmlLoader::CmlLoader(XMLHandle& handle)
     _handle = &handle;
     _scanner = 0;
     ignore_bad_valence = false;
-    valence_mode = ValenceMode::DEFAULT;
+    valence_mode = ValenceMode::BIOVIA_2009;
 }
 
 void CmlLoader::loadMolecule(Molecule& mol)
@@ -63,10 +63,11 @@ void CmlLoader::loadMolecule(Molecule& mol)
     _bmol = &mol;
     _mol = &mol;
     _qmol = 0;
-    _loadMolecule();
-
+    // Order matters: _loadMolecule() infers implicit H using the molecule's current
+    // valence model, so the mode must be set first.
     mol.setIgnoreBadValenceFlag(ignore_bad_valence);
     mol.setValenceMode(valence_mode);
+    _loadMolecule();
 }
 
 void CmlLoader::loadQueryMolecule(QueryMolecule& mol)
