+++ core/indigo-core/molecule/src/ket_document.cpp
@@ -566,6 +566,8 @@ KetDocument::KetDocument()
 }
 
 KetDocument::KetDocument(BaseMolecule& bmol)
+    : _molecules(), original_format(0), _meta_objects(rapidjson::kArrayType), _r_groups(rapidjson::kArrayType), _json_molecules(rapidjson::kArrayType),
+      _json_document()
 {
     // save molecule to ket
     std::string json;
@@ -581,6 +583,8 @@ KetDocument::KetDocument(BaseMolecule& bmol)
 }
 
 KetDocument::KetDocument(BaseReaction& breact)
+    : _molecules(), original_format(0), _meta_objects(rapidjson::kArrayType), _r_groups(rapidjson::kArrayType), _json_molecules(rapidjson::kArrayType),
+      _json_document()
 {
     // save reaction to ket
     std::string json;
