+++ core/indigo-core/molecule/src/molecule_cdxml_loader.cpp
@@ -414,7 +414,7 @@ auto MoleculeCdxmlLoader::bboxLambda(Rect2f& bbox)
 
 MoleculeCdxmlLoader::MoleculeCdxmlLoader(Scanner& scanner, bool is_binary, bool is_fragment)
     : _scanner(scanner), _is_binary(is_binary), _is_fragment(is_fragment), _has_bounding_box(false), _pmol(nullptr), _pqmol(nullptr), ignore_bad_valence(false),
-      valence_mode(ValenceMode::DEFAULT)
+      valence_mode(ValenceMode::BIOVIA_2009)
 {
 }
