+++ core/indigo-core/molecule/src/smiles_loader.cpp
@@ -41,7 +41,7 @@ SmilesLoader::SmilesLoader(Scanner& scanner) : _scanner(scanner)
     ignore_closing_bond_direction_mismatch = false;
     ignore_cistrans_errors = false;
     ignore_bad_valence = false;
-    valence_mode = ValenceMode::DEFAULT;
+    valence_mode = ValenceMode::BIOVIA_2009;
     ignore_no_chiral_flag = false;
     _mol = 0;
     _qmol = 0;
@@ -66,9 +66,11 @@ void SmilesLoader::loadMolecule(Molecule& mol)
     _qmol = 0;
     _has_atom_coordinates = false;
     mol.original_format = BaseMolecule::SMILES;
-    _loadMolecule();
+    // Order matters: _loadMolecule() infers implicit H using the molecule's current
+    // valence model, so the mode must be set first.
     mol.setIgnoreBadValenceFlag(ignore_bad_valence);
     mol.setValenceMode(valence_mode);
+    _loadMolecule();
 }
 
 void SmilesLoader::loadQueryMolecule(QueryMolecule& mol)
